requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
autopep8=2.0.4=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py311hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py311h06a4308_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
google-auth=2.23.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.58.0=pypi_0
idna=3.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py311h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.8.0=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
numpy=1.26.0=pypi_0
numpy-base=1.25.2=py311hf175353_0
oauthlib=3.2.2=pypi_0
onnx=1.14.1=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.11=h7f8727e_2
packaging=23.1=pypi_0
pandas=2.1.1=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=py311h06a4308_0
protobuf=4.24.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycodestyle=2.11.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.11_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=py311h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.11.2=pypi_0
setuptools=68.0.0=py311h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.0.2=py311_cu118
torchtriton=2.0.0=py311
torchvision=0.15.2=py311_cu118
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=py311h06a4308_0
werkzeug=2.3.7=pypi_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
